+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved Nextflow version detection in `workflowVersionToYAML()` with fallback to `NXF_VER` environment variable
 - Changed pipeline start/complete messages from `log` to `TRACE` level to reduce console noise in normal operation
 
+### Fixed
+
+- Fix versions needed for build and tests
+
 ## [0.3.1] - 2025-09-12
 
 ### Changed